# go/ecommapi/docs/swagger.yaml
definitions:
  cart.CartItemDTO:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  inventory.StockDTO:
    properties:
      increment_by:
        type: integer
      reason:
        type: string
    required:
    - increment_by
    type: object
  order.OrderDTO:
    properties:
      items:
        items:
          $ref: '#/definitions/cart.CartItemDTO'
        type: array
      shipping_city:
        type: string
      shipping_country:
        type: string
      shipping_street:
        type: string
      shipping_zip:
        type: string
    required:
    - items
    type: object
  product.ProductDTO:
    properties:
      category_id:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
    required:
    - category_id
    - name
    - price
    type: object
  user.UserLoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.UserRegisterDTO:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login as Customer or Admin
      parameters:
      - description: Login credentials
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UserLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Details of the User
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - user
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logging out as a User
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Details of the User
          schema:
            additionalProperties: true
            type: object
      summary: Logout
      tags:
      - user
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creating a new Customer
      parameters:
      - description: New User
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UserRegisterDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Details of the User
          schema:
            additionalProperties: true
            type: object
      summary: Create a User
      tags:
      - user
  /cart:
    get:
      description: Get cart items for customer
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cart Items
          schema:
            additionalProperties: true
            type: object
      summary: Get cart for customer
      tags:
      - cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Add product to the cart
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      - description: New Cart Item
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cart.CartItemDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Cart items
          schema:
            additionalProperties: true
            type: object
      summary: Add to cart
      tags:
      - cart
  /health:
    get:
      description: Check if service runs properly
      produces:
      - application/json
      responses:
        "200":
          description: Health message
          schema:
            additionalProperties: true
            type: object
      summary: Health Check
      tags:
      - health
  /inventory/:product_id:
    get:
      description: Get inventory details of a product
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Inventory details
          schema:
            additionalProperties: true
            type: object
      summary: Get product inventory
      tags:
      - inventory
    post:
      consumes:
      - application/json
      description: Increase product inventory
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      - description: Stock Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/inventory.StockDTO'
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Inventory details
          schema:
            additionalProperties: true
            type: object
      summary: Update inventory
      tags:
      - inventory
  /notifications:
    get:
      consumes:
      - application/json
      description: Get notifications of a user
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notifications
          schema:
            additionalProperties: true
            type: object
      summary: Get notifications
      tags:
      - notification
  /notifications/:id/read:
    post:
      consumes:
      - application/json
      description: Update notification read status true
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification data
          schema:
            additionalProperties: true
            type: object
      summary: Update notification status
      tags:
      - notification
  /orders:
    post:
      consumes:
      - application/json
      description: Create order from the cart
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      - description: New Order Item
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/order.OrderDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Order items
          schema:
            additionalProperties: true
            type: object
      summary: Create order
      tags:
      - order
  /products:
    get:
      description: Get products for a given category starting from a page
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product Items
          schema:
            additionalProperties: true
            type: object
      summary: Get products for category
      tags:
      - product
    post:
      consumes:
      - application/json
      description: Create a product for the given category
      parameters:
      - description: Session token
        in: header
        name: X-Session-Token
        required: true
        type: string
      - description: New Product
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/product.ProductDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            additionalProperties: true
            type: object
      summary: Create a product
      tags:
      - product
swagger: "2.0"